+++ Parser/Shared/IsExternalInit.cs
@@ -4,5 +4,5 @@
 
 namespace System.Runtime.CompilerServices;
 
-internal sealed class IsExternalInit { }
+sealed class IsExternalInit { }
 #endif